Rickshaw mostly working for leaderboards.
I've mostly got Rickshaw working. The slider is still acting
a bit funny. And it's still really ugly. However it loads
the data from a CSV based on just the CSV.

The slider is very odd. It has some serious issues and doesn't
behave exactly as I expect it to. When you zoom in close it
gets ugly quickly, and it hides points.